LayoutTests/crypto/webkitSubtle/aes-cbc-wrap-rsa-non-extractable-expected.txt

CONSOLE MESSAGE: line 32: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test wrapping an RSA key with AES-CBC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key to wrap...
PASS key.algorithm.name is 'RSASSA-PKCS1-v1_5'
Importing a key encryption key...
PASS wrappingKey.algorithm.name is 'AES-CBC'
wrapping the key...
PASS Rejected
PASS successfullyParsed is true

TEST COMPLETE
